refactor: チャットブックマーク登録ボタンを別ファイルに切り出した
